Remove fp16 support in favor of bf16
- Remove fp16 checks from trainer_phase_2.py, glue/train.py, and contrastive/trainer.py
- Update pseudo_perplexity.py to use --bf16 flag instead of --fp16
- Change hardcoded float16 to bfloat16 in run_mteb.py and model.py
- Remove fp16: false lines from all config yaml files

All modern GPUs support bf16 which provides better training stability
than fp16 while maintaining similar performance benefits.

+++ scripts/evaluation/run_mteb.py
@@ -222,7 +222,7 @@ def evaluate_mteb(cfg):
         logger.info(f"Running task: {task}")
         eval_splits = ["dev"] if task == "MSMARCO" else ["test"]
         evaluation = MTEB(tasks=[task], task_langs=["en"])
-        with torch.autocast(device_type=device, dtype=torch.float16):
+        with torch.autocast(device_type=device, dtype=torch.bfloat16):
             evaluation.run(
                 model,
                 output_folder=output_folder,
